refactor: added refactors from Init raw response controller
