+++ src/main/java/fr/insee/genesis/domain/model/surveyunit/rawdata/LunaticJsonRawDataModel.java
@@ -13,6 +13,7 @@ public record LunaticJsonRawDataModel(
     String campaignId,
     String questionnaireId,
     String interrogationId,
+    String idUE,
     Mode mode,
     LunaticJsonRawData data,
     LocalDateTime recordDate,
