+++ src/main/resources/application.properties
@@ -9,6 +9,7 @@ spring.profiles.active=local
 fr.insee.genesis.authentication = NONE
     #NONE = No authentication required (By default for LOCAL and DEV)
     #OIDC = Keycloack authentication (PROD, override by Puppet/ConfigMap)
+logging.level.org.springframework.security=DEBUG
 
 #--------------------------------------------------------------------------
 # Configuration for springdoc / swagger
