+++ src/test/java/fr/insee/genesis/RunCucumberTest.java
@@ -1,3 +1,5 @@
+package fr.insee.genesis;
+
 import io.cucumber.junit.Cucumber;
 import io.cucumber.junit.CucumberOptions;
 import org.junit.runner.RunWith;
